requirements.txt
web3
# Required for pushover notifications
python-pushover